Restructure into spanner interactions and low level grpc
